Fix least similar movies index selection
This is a bit pedantic, but the "least" similar example used the same 1:11 range as the most similar (rather than 0:10, likely due to copy-pasting from the above cell where that range makes more sense). Probably doesn't _really _matter, but I figured this series was quite helpful to me, so I'd like to pay it forward if I can and avoid the half-second of confusion for the next person who looks at this example.